+++ CHANGELOG.md
@@ -93,6 +93,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   inline errors, and more. Please [see the documentation](/docs/config.md#json-schema-support)
   for more on this.
 
+* Add a config option `ui.show-cryptographic-signatures`. When set to `"true"`, the
+  builtin templates will show signataure information if available. The signature display
+  can be customized using `format_detailed_cryptographic_signature(signature)` and
+  `format_short_cryptographic_signature(signature)`.
+
 ### Fixed bugs
 
 * Fixed diff selection by external tools with `jj split`/`commit -i FILESETS`.